Manually merging Matrix/Align code from Learnosity/mathquill
KUDOS TO:
https://github.com/alexf101
https://github.com/eoghanmcilwaine

I was unable to successfully git  merge the branch from Learnosity                                                           that contained code for matrices and alignment. The branch I                                                            was intending to pull was:
https://github.com/Learnosity/mathquill/tree/align-alternative.
Thus, I manually copied the changes across.

I have to sincerely thanks Alex and Eoghan for their contributions
and take no credit for the aforementioned work.